+++ test/formatters/installers/hurlfmt.bash
@@ -1,0 +1,2 @@
+sudo apt-add-repository -y ppa:lepapareil/hurl
+sudo apt-get install hurl