+++ test/formatters/installers/csharpier.bash
@@ -1,0 +1,2 @@
+apt-get install dotnet-sdk-8.0
+dotnet tool install -g csharpier