+++ test/formatters/samplecode/csharpier/in.cs
@@ -1,0 +1,3 @@
+  void
+    Foo(){   return   666 ;// comment of the beast
+}