+++ apheleia-utils.el
@@ -48,6 +48,7 @@ always returns nil to defer to the formatter."
               (js2-mode 'js2-basic-offset)
               (js3-mode 'js3-indent-level)
               (json-mode 'js-indent-level)
+              (js-json-mode 'js-indent-level)
               (json-ts-mode 'json-ts-mode-indent-offset)
               (nxml-mode 'nxml-child-indent)
               (robot-mode 'robot-mode-basic-offset)