+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog].
 
 ## Unreleased
 ### Bugs fixed
+* aphelia-formatters-indent did not handle yaml-ts-mode; added with
+  same behaviour as yaml-mode
 * A formatter that adds indentation while point is at the end of the
   line would sometimes leave point at the wrong position ([#362]).
 ### Formatters