+++ CHANGELOG.md
@@ -10,8 +10,8 @@ The format is based on [Keep a Changelog].
 * A formatter that adds indentation while point is at the end of the
   line would sometimes leave point at the wrong position ([#362]).
 ### Formatters
-* [hurlfmt](https://hurl.dev)
-  for hurl files.
+* [hurlfmt](https://hurl.dev) for hurl files.
+* [csharpier](https://github.com/belav/csharpier) for C#.
 
 ## 4.4.1 (released 2025-05-13)
 ### Enhancements